requirements.txt
ffmpeg-python==0.2.0
future==1.0.0
pytubefix==6.11.0
PyYAML==6.0.2
